backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
boto3==1.36.2
python-dotenv==1.0.1
pydantic==2.10.5
mangum==0.19.0
python-multipart==0.0.20
